backend/requirements.txt
azure-functions
azure-search-documents
azure-ai-formrecognizer
azure-storage-blob
openai
requests
python-dotenv
